Synk oppgave må gjøres som task for å få rettigheter til å kalle oppgave-systemet med maskin-token.
